requirements.txt
python==3.10.2
python-telegram-bot==12.7
